kconfig: "output yaml" renders boolean N as a comment
I found that the "disable_skipped_hosts" Kconfig setting was not
working.

Using "output yaml" with a boolean behaves somewhat counter-
intuitively. When the user specifies "Yes, please set this", the
variable appears in .extra_vars_auto.yaml with a value of True.
When the user specifies "No, please do not set this" the variable
does not appear in .extra_vars_auto.yaml at all.

A more convenient and sensible behavior would be for "output yaml"
to output false Kconfig settings with an explicit "variable: False".
That way the Kconfig setting carries through as an Ansible variable
no matter if the setting is Y or N.

Suggested-by: Daniel Gomez <da.gomez@kernel.org>
Acked-by: Daniel Gomez <da.gomez@samsung.com>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>