train.py: add support for AdamW
So we actually were leveraging the Adam optimizer since the digital
ocean inspired implementation, not SGD. That's fine we have the data,
but lets move on.

So what we are learning is we want to be optimzizer flexible and, we
want many options. So let's just do that.

These days favored optimizer is AdamW, so add support for it. And we'll stick
to SGD by default but we'll also keep Adam as an option. Since Adam was
already used we are updating our graphs to reflect SGD vanilla results.

Then extend our Makefile and scripts to be dynamic we can easily
reproduce our findings and update graphs. When in doubt just use:

  make help

To reproduce all findings run:

  make

To update our graphs:

  make update-graphs

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>