optimizers: add an adamwprune "base"
The idea behind pruning using adam states should not be confused with
arbitrary adam optimizer enhancements. And so evaluating admamwprune
should be adaptable to any base adam optimizer:

  * adam
  * adamw
  * adamwadv (list of advanced features in R&D literature)
  * adamwspam (with spam support)

It seems on resnet18 the latest and greatest does not necessarily give
us the best accuracy, so the simple adam optimizer will suffice for some
tests. So just make this explicit so that we can clarify that adamwprune
is adaptable to any base adam optimizer.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>